+++ include/pkgi.h
@@ -158,6 +158,8 @@ void pkgi_draw_fill_rect(int x, int y, int w, int h, uint32_t color);
 void pkgi_draw_fill_rect_z(int x, int y, int z, int w, int h, uint32_t color);
 void pkgi_draw_text(int x, int y, uint32_t color, const char* text);
 void pkgi_draw_text_z(int x, int y, int z, uint32_t color, const char* text);
+void pkgi_draw_text_ttf(int x, int y, int z, uint32_t color, const char* text);
 int pkgi_text_width(const char* text);
+int pkgi_text_width_ttf(const char* text);
 int pkgi_text_height(const char* text);