+++ include/pkgi_dialog.h
@@ -7,7 +7,7 @@ void pkgi_dialog_init(void);
 int pkgi_dialog_is_open(void);
 int pkgi_dialog_is_cancelled(void);
 void pkgi_dialog_allow_close(int allow);
-void pkgi_dialog_message(const char* text);
+void pkgi_dialog_message(const char* title, const char* text);
 void pkgi_dialog_error(const char* text);
 void pkgi_dialog_details(const char* title, const char* text, const char* extra);