+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the `pkgi-ps3` project will be documented in this file. T
 
 ## [Unreleased]()
 
+### Added
+
+* Added TTF fonts to support Japanese characters
+
 ## [v1.0.8](https://github.com/bucanero/pkgi-ps3/releases/tag/v1.0.8) - 2019-12-19
 
 ### Added